complete refactor null check
