+++ System/src/Checks/ExceptionActivator.cs
@@ -6,17 +6,25 @@ internal static class ExceptionActivator
 {
 	internal static T CreateGenericInstance<T>([InvokerParameterName] string paramName)
 		where T : Exception
-		=> (Activator.CreateInstance(typeof(T), paramName) as T)!;
+	{
+		return (Activator.CreateInstance(typeof(T), paramName) as T)!;
+	}
 
 	internal static T CreateGenericInstance<T>([InvokerParameterName] string paramName, string message)
 		where T : Exception
-		=> (Activator.CreateInstance(typeof(T), paramName, message) as T)!;
+	{
+		return (Activator.CreateInstance(typeof(T), paramName, message) as T)!;
+	}
 
 	internal static T CreateArgumentInstance<T>([InvokerParameterName] string paramName)
 		where T : ArgumentException
-		=> (Activator.CreateInstance(typeof(T), paramName) as T)!;
+	{
+		return (Activator.CreateInstance(typeof(T), paramName) as T)!;
+	}
 
 	internal static T CreateArgumentInstance<T>([InvokerParameterName] string paramName, string message)
 		where T : ArgumentException
-		=> (Activator.CreateInstance(typeof(T), paramName, message) as T)!;
+	{
+		return (Activator.CreateInstance(typeof(T), paramName, message) as T)!;
+	}
 }
