+++ System/src/Checks/ThrowIfEmptyStringExtensions.cs
@@ -10,26 +10,36 @@ public static class ThrowIfEmptyStringExtensions
 {
 	// Throw if value string is empty
 	public static string ThrowIfEmpty(this string? value)
-		=> value.ThrowIfNull().ThrowIfEmpty<ArgumentEmptyException>();
+	{
+		return value.ThrowIfNull().ThrowIfEmpty<ArgumentEmptyException>();
+	}
 
 	public static string ThrowIfEmpty(this string? value, string message)
-		=> value.ThrowIfNull().ThrowIfEmpty<ArgumentEmptyException>(message);
+	{
+		return value.ThrowIfNull().ThrowIfEmpty<ArgumentEmptyException>(message);
+	}
 
 	public static string ThrowIfEmpty<T>(this string? value)
 		where T : ArgumentException
-		=> value.ThrowIfNull<T>().IsEmpty()
-			   ? throw ExceptionActivator.CreateArgumentInstance<T>(nameof(value))
-			   : value!;
+	{
+		return value.ThrowIfNull<T>().IsEmpty()
+			       ? throw ExceptionActivator.CreateArgumentInstance<T>(nameof(value))
+			       : value!;
+	}
 
 	public static string ThrowIfEmpty<T>(this string? value, string message)
 		where T : ArgumentException
-		=> value.ThrowIfNull<T>().IsEmpty()
-			   ? throw ExceptionActivator.CreateArgumentInstance<T>(nameof(value), message)
-			   : value!;
+	{
+		return value.ThrowIfNull<T>().IsEmpty()
+			       ? throw ExceptionActivator.CreateArgumentInstance<T>(nameof(value), message)
+			       : value!;
+	}
 
 	public static string ThrowIfEmpty<T>(this string? value, string message, [InvokerParameterName] string paramName)
 		where T : ArgumentException
-		=> value.ThrowIfNull<T>().IsEmpty()
-			   ? throw ExceptionActivator.CreateArgumentInstance<T>(paramName, message)
-			   : value!;
+	{
+		return value.ThrowIfNull<T>().IsEmpty()
+			       ? throw ExceptionActivator.CreateArgumentInstance<T>(paramName, message)
+			       : value!;
+	}
 }
